[refactor] 기존 코드 리팩토링
