+++ public/common/elasticSearch/createIndex.js
