+++ public/common/elasticSearch/esIndexObject.js
