+++ public/common/elasticSearch/insertData.js
@@ -1,8 +1,11 @@
-const { Client } = require("@elastic/elasticsearch"); // ElasticSearch 사용을 위한
+import { Client } from "@elastic/elasticsearch";
+
 const esClient = new Client({ node: "http://localhost:9200" });
 
+const indexName = "themes";
+
 const insertData = async (data) => {
-  const body = data.flatMap((doc) => [{ index: { _index: "themes" } }, doc]);
+  const body = data.flatMap((doc) => [{ index: { _index: "indexName" } }, doc]);
   await esClient.bulk({ refresh: true, body });
 };
